requirements.txt
python-telegram-bot==20.7
openai==1.12.0
feedparser==6.0.10
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.1
